Remove stale inspectors and fields
