+++ apps/annotation/filters.py
@@ -1,6 +1,10 @@
+from distutils.util import strtobool
+
 import coreapi
 import coreschema
-import django_filters.fields
+import django_filters
+from django.forms import CharField
+from django_filters.constants import EMPTY_VALUES
 from drf_yasg import openapi
 from rest_framework.filters import BaseFilterBackend
 
@@ -74,3 +78,37 @@ class ListORFilter(django_filters.Filter):
         if value:
             value_list = value.split(',')
         return super(ListORFilter, self).filter(qs, value_list)
+
+
+class NullBooleanField(CharField):
+    def to_python(self, value):
+        if value is None:
+            return None
+        try:
+            return bool(strtobool(value))
+        except ValueError:
+            return None
+
+    def validate(self, value):
+        pass
+
+
+class BooleanFilter(django_filters.Filter):
+    field_class = NullBooleanField
+
+
+class RequestUserBooleanFilter(django_filters.Filter):
+    field_class = NullBooleanField
+
+    def get_method_based_on_value(self, qs, value):
+        return qs.filter if value else qs.exclude
+
+    def filter(self, qs, value):
+        if value in EMPTY_VALUES:
+            return qs
+        if self.distinct:
+            qs = qs.distinct()
+        lookup = '%s__%s' % (self.field_name, self.lookup_expr)
+        qs = self.get_method_based_on_value(qs, value)(**{lookup: self.parent.request.user})
+        return qs
+
